Moves sorter to its own file
