+++ frontend/src/Types/SortOrder.ts
@@ -1,0 +1,3 @@
+type SortOrder = 'asc' | 'desc';
+
+export default SortOrder;
